requirements.txt
fastapi==0.109.2
langchain==0.2.7
langchain-openai==0.1.15
python-dotenv==1.0.1
uvicorn==0.27.0